+++ web/src/lib/components/pages/profile/profile.css
@@ -1,0 +1,41 @@
+#profile-container h2 {
+    padding: 0.8rem 0;;
+}
+
+#profile-container hr {
+    margin-top: 1.5rem;
+}
+
+#profile-container select {
+    margin-bottom: 0.5rem;
+}
+
+#profile-container #save-demos {
+    margin-top: 1rem;
+    width: 100%;
+    height: var(--input-height);
+}
+
+#profile-container .unsaved {
+    background-color: yellow;
+}
+
+#opt-in {
+    margin-top: 1rem;
+}
+
+input[type="checkbox"]{
+    width: .8rem;
+    height: .8rem;
+    margin: .1rem .1rem .1rem .1rem;
+}
+
+label[for="send-emails"] {
+    font-size: 1rem;
+}
+
+#email-explanation {
+    padding-top: 0.5rem;
+    padding-left: 1rem;
+    font-size: 0.8rem;
+}
