+++ web/src/lib/components/pages/record/audio-ios.ts
@@ -1,7 +1,7 @@
 import ERROR_MSG from '../../../../error-msg';
 import { isNativeIOS } from '../../../utility';
 import { AudioInfo } from './audio-web';
-import confirm from '../../confirm';
+import confirm from '../../../confirm/confirm';
 
 const LEVELS_THROTTLE = 50;
