+++ web/src/lib/components/pages/record/audio-web.ts
@@ -1,6 +1,6 @@
 import ERROR_MSG from '../../../../error-msg';
 import { isNativeIOS } from '../../../utility';
-import confirm from '../../confirm';
+import confirm from '../../../confirm/confirm';
 
 const AUDIO_TYPE = 'audio/ogg; codecs=opus';
