+++ web/src/lib/confirm/confirm.css
@@ -1,0 +1,46 @@
+.confirm-modal {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    z-index: var(--top-z-index);
+    pointer-events: none;
+    opacity: 0;
+    transition: opacity 0.2s;
+    background: rgba(0, 0, 0, 0.8);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.confirm-modal label {
+    display: block;
+    margin-bottom: 1rem;
+    font-size: larger;
+}
+
+.confirm-modal button {
+    padding: 0.5rem 1rem;
+    margin: 0.5rem;
+    min-width: 10rem;
+}
+
+.confirm-modal.visible {
+    pointer-events: all;
+    opacity: 1;
+}
+
+.confirm-modal .main-section {
+    display: flex;
+    flex-direction: column;
+    margin: 1rem;
+    padding: 1rem;
+    background: var(--base-background-color);
+    width: 100%;
+    max-width: 30rem;
+}
+
+#confirm-confirm {
+    background-color: var(--color-help);
+}
